+++ .gitignore
@@ -7,6 +7,7 @@ venv
 
 build/
 output/
+dist/
 
 eth2.0-spec-tests/
 
@@ -14,8 +15,8 @@ eth2.0-spec-tests/
 .mypy_cache
 
 # Dynamically built from Markdown spec
-tests/core/pyspec/eth2spec/phase0/spec.py
-tests/core/pyspec/eth2spec/phase1/spec.py
+tests/core/pyspec/eth2spec/phase0/
+tests/core/pyspec/eth2spec/phase1/
 
 # coverage reports
 .htmlcov
@@ -24,5 +25,6 @@ tests/core/pyspec/eth2spec/phase1/spec.py
 
 # local CI testing output
 tests/core/pyspec/test-reports
+tests/core/pyspec/eth2spec/test_results.xml
 
 *.egg-info
