+++ setup.py
@@ -1,0 +1,465 @@
+from setuptools import setup, find_packages, Command
+from setuptools.command.build_py import build_py
+from distutils import dir_util
+from distutils.util import convert_path
+import os
+import re
+from typing import Dict, NamedTuple, List
+
+FUNCTION_REGEX = r'^def [\w_]*'
+
+
+class SpecObject(NamedTuple):
+    functions: Dict[str, str]
+    custom_types: Dict[str, str]
+    constants: Dict[str, str]
+    ssz_objects: Dict[str, str]
+
+
+def get_spec(file_name: str) -> SpecObject:
+    """
+    Takes in the file name of a spec.md file, opens it and returns a parsed spec object.
+
+    Note: This function makes heavy use of the inherent ordering of dicts,
+    if this is not supported by your python version, it will not work.
+    """
+    pulling_from = None  # line number of start of latest object
+    current_name = None  # most recent section title
+    functions: Dict[str, str] = {}
+    constants: Dict[str, str] = {}
+    ssz_objects: Dict[str, str] = {}
+    function_matcher = re.compile(FUNCTION_REGEX)
+    is_ssz = False
+    custom_types: Dict[str, str] = {}
+    for linenum, line in enumerate(open(file_name).readlines()):
+        line = line.rstrip()
+        if pulling_from is None and len(line) > 0 and line[0] == '#' and line[-1] == '`':
+            current_name = line[line[:-1].rfind('`') + 1: -1]
+        if line[:9] == '```python':
+            assert pulling_from is None
+            pulling_from = linenum + 1
+        elif line[:3] == '```':
+            pulling_from = None
+        else:
+            # Handle function definitions & ssz_objects
+            if pulling_from is not None:
+                # SSZ Object
+                if len(line) > 18 and line[:6] == 'class ' and line[-12:] == '(Container):':
+                    name = line[6:-12]
+                    # Check consistency with markdown header
+                    assert name == current_name
+                    is_ssz = True
+                # function definition
+                elif function_matcher.match(line) is not None:
+                    current_name = function_matcher.match(line).group(0)
+                    is_ssz = False
+                if is_ssz:
+                    ssz_objects[current_name] = ssz_objects.get(current_name, '') + line + '\n'
+                else:
+                    functions[current_name] = functions.get(current_name, '') + line + '\n'
+            # Handle constant and custom types table entries
+            elif pulling_from is None and len(line) > 0 and line[0] == '|':
+                row = line[1:].split('|')
+                if len(row) >= 2:
+                    for i in range(2):
+                        row[i] = row[i].strip().strip('`')
+                        if '`' in row[i]:
+                            row[i] = row[i][:row[i].find('`')]
+                    is_constant_def = True
+                    if row[0][0] not in 'ABCDEFGHIJKLMNOPQRSTUVWXYZ_':
+                        is_constant_def = False
+                    for c in row[0]:
+                        if c not in 'ABCDEFGHIJKLMNOPQRSTUVWXYZ_0123456789':
+                            is_constant_def = False
+                    if is_constant_def:
+                        constants[row[0]] = row[1].replace('**TBD**', '2**32')
+                    elif row[1].startswith('uint') or row[1].startswith('Bytes'):
+                        custom_types[row[0]] = row[1]
+    return SpecObject(functions, custom_types, constants, ssz_objects)
+
+
+CONFIG_LOADER = '''
+apply_constants_config(globals())
+'''
+
+PHASE0_IMPORTS = '''from eth2spec.config.config_util import apply_constants_config
+from typing import (
+    Any, Callable, Dict, Set, Sequence, Tuple, Optional, TypeVar
+)
+
+from dataclasses import (
+    dataclass,
+    field,
+)
+
+from eth2spec.utils.ssz.ssz_impl import hash_tree_root
+from eth2spec.utils.ssz.ssz_typing import (
+    View, boolean, Container, List, Vector, uint64,
+    Bytes1, Bytes4, Bytes8, Bytes32, Bytes48, Bytes96, Bitlist, Bitvector,
+)
+from eth2spec.utils import bls
+
+from eth2spec.utils.hash_function import hash
+
+SSZObject = TypeVar('SSZObject', bound=View)
+'''
+PHASE1_IMPORTS = '''from eth2spec.phase0 import spec as phase0
+from eth2spec.config.config_util import apply_constants_config
+from typing import (
+    Any, Dict, Set, Sequence, NewType, Tuple, TypeVar, Callable, Optional
+)
+
+from dataclasses import (
+    dataclass,
+    field,
+)
+
+from eth2spec.utils.ssz.ssz_impl import hash_tree_root
+from eth2spec.utils.ssz.ssz_typing import (
+    View, boolean, Container, List, Vector, uint64, uint8, bit,
+    ByteVector, ByteList, Bytes1, Bytes4, Bytes8, Bytes32, Bytes48, Bytes96, Bitlist, Bitvector,
+)
+from eth2spec.utils import bls
+
+from eth2spec.utils.hash_function import hash
+
+# Whenever phase 1 is loaded, make sure we have the latest phase0
+from importlib import reload
+reload(phase0)
+
+
+SSZVariableName = str
+GeneralizedIndex = NewType('GeneralizedIndex', int)
+SSZObject = TypeVar('SSZObject', bound=View)
+'''
+SUNDRY_CONSTANTS_FUNCTIONS = '''
+def ceillog2(x: uint64) -> int:
+    return (x - 1).bit_length()
+'''
+SUNDRY_FUNCTIONS = '''
+# Monkey patch hash cache
+_hash = hash
+hash_cache: Dict[bytes, Bytes32] = {}
+
+
+def get_eth1_data(distance: uint64) -> Bytes32:
+    return hash(distance)
+
+
+def hash(x: bytes) -> Bytes32:  # type: ignore
+    if x not in hash_cache:
+        hash_cache[x] = Bytes32(_hash(x))
+    return hash_cache[x]
+
+
+def cache_this(key_fn, value_fn):  # type: ignore
+    cache_dict = {}  # type: ignore
+
+    def wrapper(*args, **kw):  # type: ignore
+        key = key_fn(*args, **kw)
+        nonlocal cache_dict
+        if key not in cache_dict:
+            cache_dict[key] = value_fn(*args, **kw)
+        return cache_dict[key]
+    return wrapper
+
+
+get_base_reward = cache_this(
+    lambda state, index: (state.validators.hash_tree_root(), state.slot),
+    get_base_reward)
+
+get_committee_count_at_slot = cache_this(
+    lambda state, epoch: (state.validators.hash_tree_root(), epoch),
+    get_committee_count_at_slot)
+
+get_active_validator_indices = cache_this(
+    lambda state, epoch: (state.validators.hash_tree_root(), epoch),
+    get_active_validator_indices)
+
+get_beacon_committee = cache_this(
+    lambda state, slot, index: (state.validators.hash_tree_root(), state.randao_mixes.hash_tree_root(), slot, index),
+    get_beacon_committee)'''
+
+
+def objects_to_spec(spec_object: SpecObject, imports: str, version: str) -> str:
+    """
+    Given all the objects that constitute a spec, combine them into a single pyfile.
+    """
+    new_type_definitions = (
+        '\n\n'.join(
+            [
+                f"class {key}({value}):\n    pass\n"
+                for key, value in spec_object.custom_types.items()
+            ]
+        )
+    )
+    for k in list(spec_object.functions):
+        if "ceillog2" in k:
+            del spec_object.functions[k]
+    functions_spec = '\n\n'.join(spec_object.functions.values())
+    for k in list(spec_object.constants.keys()):
+        if k == "BLS12_381_Q":
+            spec_object.constants[k] += "  # noqa: E501"
+    constants_spec = '\n'.join(map(lambda x: '%s = %s' % (x, spec_object.constants[x]), spec_object.constants))
+    ssz_objects_instantiation_spec = '\n\n'.join(spec_object.ssz_objects.values())
+    spec = (
+            imports
+            + '\n\n' + f"version = \'{version}\'\n"
+            + '\n\n' + new_type_definitions
+            + '\n' + SUNDRY_CONSTANTS_FUNCTIONS
+            + '\n\n' + constants_spec
+            + '\n\n' + CONFIG_LOADER
+            + '\n\n' + ssz_objects_instantiation_spec
+            + '\n\n' + functions_spec
+            + '\n' + SUNDRY_FUNCTIONS
+            + '\n'
+    )
+    return spec
+
+
+def combine_functions(old_functions: Dict[str, str], new_functions: Dict[str, str]) -> Dict[str, str]:
+    for key, value in new_functions.items():
+        old_functions[key] = value
+    return old_functions
+
+
+def combine_constants(old_constants: Dict[str, str], new_constants: Dict[str, str]) -> Dict[str, str]:
+    for key, value in new_constants.items():
+        old_constants[key] = value
+    return old_constants
+
+
+ignored_dependencies = [
+    'bit', 'boolean', 'Vector', 'List', 'Container', 'BLSPubkey', 'BLSSignature',
+    'Bytes1', 'Bytes4', 'Bytes32', 'Bytes48', 'Bytes96', 'Bitlist', 'Bitvector',
+    'uint8', 'uint16', 'uint32', 'uint64', 'uint128', 'uint256',
+    'bytes', 'byte', 'ByteList', 'ByteVector'
+]
+
+
+def dependency_order_ssz_objects(objects: Dict[str, str], custom_types: Dict[str, str]) -> None:
+    """
+    Determines which SSZ Object is dependent on which other and orders them appropriately
+    """
+    items = list(objects.items())
+    for key, value in items:
+        dependencies = []
+        for line in value.split('\n'):
+            if not re.match(r'\s+\w+: .+', line):
+                continue  # skip whitespace etc.
+            line = line[line.index(':') + 1:]  # strip of field name
+            if '#' in line:
+                line = line[:line.index('#')]  # strip of comment
+            dependencies.extend(re.findall(r'(\w+)', line))  # catch all legible words, potential dependencies
+        dependencies = filter(lambda x: '_' not in x and x.upper() != x, dependencies)  # filter out constants
+        dependencies = filter(lambda x: x not in ignored_dependencies, dependencies)
+        dependencies = filter(lambda x: x not in custom_types, dependencies)
+        for dep in dependencies:
+            key_list = list(objects.keys())
+            for item in [dep, key] + key_list[key_list.index(dep)+1:]:
+                objects[item] = objects.pop(item)
+
+
+def combine_ssz_objects(old_objects: Dict[str, str], new_objects: Dict[str, str], custom_types) -> Dict[str, str]:
+    """
+    Takes in old spec and new spec ssz objects, combines them,
+    and returns the newer versions of the objects in dependency order.
+    """
+    for key, value in new_objects.items():
+        old_objects[key] = value
+    return old_objects
+
+
+def combine_spec_objects(spec0: SpecObject, spec1: SpecObject) -> SpecObject:
+    """
+    Takes in two spec variants (as tuples of their objects) and combines them using the appropriate combiner function.
+    """
+    functions0, custom_types0, constants0, ssz_objects0 = spec0
+    functions1, custom_types1, constants1, ssz_objects1 = spec1
+    functions = combine_functions(functions0, functions1)
+    custom_types = combine_constants(custom_types0, custom_types1)
+    constants = combine_constants(constants0, constants1)
+    ssz_objects = combine_ssz_objects(ssz_objects0, ssz_objects1, custom_types)
+    return SpecObject(functions, custom_types, constants, ssz_objects)
+
+
+def dependency_order_spec(objs: SpecObject):
+    functions, custom_types, constants, ssz_objects = objs
+    dependency_order_ssz_objects(ssz_objects, custom_types)
+
+
+version_imports = {
+    'phase0': PHASE0_IMPORTS,
+    'phase1': PHASE1_IMPORTS,
+}
+
+
+def build_spec(version: str, source_files: List[str]) -> str:
+    all_spescs = [get_spec(spec) for spec in source_files]
+
+    spec_object = all_spescs[0]
+    for value in all_spescs[1:]:
+        spec_object = combine_spec_objects(spec_object, value)
+
+    dependency_order_spec(spec_object)
+
+    return objects_to_spec(spec_object, version_imports[version], version)
+
+
+class PySpecCommand(Command):
+    """Convert spec markdown files to a spec python file"""
+
+    description = "Convert spec markdown files to a spec python file"
+
+    spec_version: str
+    md_doc_paths: str
+    parsed_md_doc_paths: List[str]
+    out_dir: str
+
+    # The format is (long option, short option, description).
+    user_options = [
+        ('spec-version=', None, "Spec version to tag build with. Used to select md-docs defaults."),
+        ('md-doc-paths=', None, "List of paths of markdown files to build spec with"),
+        ('out-dir=', None, "Output directory to write spec package to")
+    ]
+
+    def initialize_options(self):
+        """Set default values for options."""
+        # Each user option must be listed here with their default value.
+        self.spec_version = 'phase0'
+        self.md_doc_paths = ''
+        self.out_dir = 'pyspec_output'
+
+    def finalize_options(self):
+        """Post-process options."""
+        if len(self.md_doc_paths) == 0:
+            if self.spec_version == "phase0":
+                self.md_doc_paths = """
+                    specs/phase0/beacon-chain.md
+                    specs/phase0/fork-choice.md
+                    specs/phase0/validator.md
+                """
+            elif self.spec_version == "phase1":
+                self.md_doc_paths = """
+                    specs/phase0/beacon-chain.md
+                    specs/phase0/fork-choice.md
+                    specs/phase1/custody-game.md
+                    specs/phase1/beacon-chain.md
+                    specs/phase1/fraud-proofs.md
+                    specs/phase1/fork-choice.md
+                    specs/phase1/phase1-fork.md
+                """
+            else:
+                raise Exception('no markdown files specified, and spec version "%s" is unknown', self.spec_version)
+            print("no paths were specified, using default markdown file paths for pyspec build (spec version: %s)" % self.spec_version)
+
+        self.parsed_md_doc_paths = self.md_doc_paths.split()
+
+        for filename in self.parsed_md_doc_paths:
+            if not os.path.exists(filename):
+                raise Exception('Pyspec markdown input file "%s" does not exist.' % filename)
+
+    def run(self):
+        spec_str = build_spec(self.spec_version, self.parsed_md_doc_paths)
+        if self.dry_run:
+            self.announce('dry run successfully prepared contents for spec.'
+                          f' out dir: "{self.out_dir}", spec version: "{self.spec_version}"')
+            self.debug_print(spec_str)
+        else:
+            dir_util.mkpath(self.out_dir)
+            with open(os.path.join(self.out_dir, 'spec.py'), 'w') as out:
+                out.write(spec_str)
+            with open(os.path.join(self.out_dir, '__init__.py'), 'w') as out:
+                out.write("")
+
+
+class BuildPyCommand(build_py):
+    """Customize the build command to run the spec-builder on setup.py build"""
+
+    def initialize_options(self):
+        super(BuildPyCommand, self).initialize_options()
+
+    def run_pyspec_cmd(self, spec_version: str, **opts):
+        cmd_obj: PySpecCommand = self.distribution.reinitialize_command("pyspec")
+        cmd_obj.spec_version = spec_version
+        cmd_obj.out_dir = os.path.join(self.build_lib, 'eth2spec', spec_version)
+        for k, v in opts.items():
+            setattr(cmd_obj, k, v)
+        self.run_command('pyspec')
+
+    def run(self):
+        self.run_pyspec_cmd(spec_version="phase0")
+        self.run_pyspec_cmd(spec_version="phase1")
+
+        super(BuildPyCommand, self).run()
+
+
+class PyspecDevCommand(Command):
+    """Build the markdown files in-place to their source location for testing."""
+    description = "Build the markdown files in-place to their source location for testing."
+    user_options = []
+
+    def initialize_options(self):
+        pass
+
+    def finalize_options(self):
+        pass
+
+    def run_pyspec_cmd(self, spec_version: str, **opts):
+        cmd_obj: PySpecCommand = self.distribution.reinitialize_command("pyspec")
+        cmd_obj.spec_version = spec_version
+        eth2spec_dir = convert_path(self.distribution.package_dir['eth2spec'])
+        cmd_obj.out_dir = os.path.join(eth2spec_dir, spec_version)
+        for k, v in opts.items():
+            setattr(cmd_obj, k, v)
+        self.run_command('pyspec')
+
+    def run(self):
+        print("running build_py command")
+        self.run_pyspec_cmd(spec_version="phase0", package_inplace=False)
+        self.run_pyspec_cmd(spec_version="phase1", package_inplace=False)
+
+
+commands = {
+    'pyspec': PySpecCommand,
+    'build_py': BuildPyCommand,
+    'pyspecdev': PyspecDevCommand,
+}
+
+with open("README.md", "rt", encoding="utf8") as f:
+    readme = f.read()
+
+setup(
+    name='eth2spec',
+    description="Eth2 spec, provided as Python package for tooling and testing",
+    long_description=readme,
+    long_description_content_type="text/markdown",
+    author="ethereum",
+    url="https://github.com/ethereum/eth2.0-specs",
+    include_package_data=False,
+    package_data={'configs': ['*.yaml'],
+                  'specs': ['**/*.md']},
+    package_dir={
+        "eth2spec": "tests/core/pyspec/eth2spec",
+        "configs": "configs",
+        "specs": "specs"
+    },
+    packages=find_packages(where='tests/core/pyspec') + ['configs', 'specs'],
+    py_modules=["eth2spec"],
+    cmdclass=commands,
+    python_requires=">=3.8, <4",
+    tests_require=[],  # avoid old style tests require. Enable explicit (re-)installs, e.g. `pip install .[testing]`
+    extras_require={
+        "testing": ["pytest>=4.4", "pytest-cov", "pytest-xdist"],
+        "linting": ["flake8==3.7.7", "mypy==0.750"],
+    },
+    install_requires=[
+        "eth-utils>=1.3.0,<2",
+        "eth-typing>=2.1.0,<3.0.0",
+        "pycryptodome==3.9.4",
+        "py_ecc==2.0.0",
+        "dataclasses==0.6",
+        "remerkleable==0.1.10",
+        "ruamel.yaml==0.16.5"
+    ]
+)
