--- tests/core/config_helpers/README.md
@@ -1,19 +1,0 @@
-# Eth2 config helpers
-
-`preset_loader`: A util to load config-presets with.
-See [Configs documentation](../../../configs/README.md).
-
-Usage:
-
-```python
-configs_path = 'configs/'
-
-...
-
-import preset_loader
-from eth2spec.phase0 import spec
-my_presets = preset_loader.load_presets(configs_path, 'mainnet')
-spec.apply_constants_preset(my_presets)
-```
-
-WARNING: this overwrites globals, make sure to prevent accidental collisions with other usage of the same imported specs package.
