--- tests/core/config_helpers/requirements.txt
@@ -1,1 +1,0 @@
-ruamel.yaml==0.16.5
