--- tests/core/pyspec/requirements.txt
@@ -1,6 +1,0 @@
-eth-utils>=1.3.0,<2
-eth-typing>=2.1.0,<3.0.0
-pycryptodome==3.9.4
-py_ecc==2.0.0
-dataclasses==0.6
-remerkleable==0.1.10
