+++ tests/generators/operations/main.py
@@ -11,7 +11,8 @@ from eth2spec.test.phase_0.block_processing import (
 
 from gen_base import gen_runner, gen_typing
 from gen_from_tests.gen import generate_from_tests
-from preset_loader import loader
+from importlib import reload
+from eth2spec.config import config_util
 from eth2spec.phase0 import spec as spec_phase0
 from eth2spec.phase1 import spec as spec_phase1
 
@@ -19,9 +20,9 @@ from eth2spec.phase1 import spec as spec_phase1
 def create_provider(handler_name: str, tests_src, config_name: str) -> gen_typing.TestProvider:
 
     def prepare_fn(configs_path: str) -> str:
-        presets = loader.load_presets(configs_path, config_name)
-        spec_phase0.apply_constants_preset(presets)
-        spec_phase1.apply_constants_preset(presets)
+        config_util.prepare_config(configs_path, config_name)
+        reload(spec_phase0)
+        reload(spec_phase1)
         return config_name
 
     def cases_fn() -> Iterable[gen_typing.TestCase]:
